Now working also in VS6. Fixed warnings and small glitches.
git-svn-id: svn://falconpl.org/falcon/core/trunk@288 78e5f83a-d9f7-4e86-b854-dad15b5eb37a

+++ engine/core_ext.cpp
@@ -1717,7 +1717,7 @@ FALCON_FUNC  Iterator_value( ::Falcon::VMachine *vm )
             if( subst != 0 )
             {
                if ( subst->isOrdinal() )
-                  porigin->asMemBuf()->set( p, subst->forceInteger() );
+                  porigin->asMemBuf()->set( p, (uint32) subst->forceInteger() );
                else
                   break; // only numbers allowed
             }
@@ -2112,10 +2112,10 @@ FALCON_FUNC Make_MemBuf( ::Falcon::VMachine *vm )
    MemBuf *mb = 0;
    switch( wordSize )
    {
-      case 1: mb = new MemBuf_1( vm, size ); break;
-      case 2: mb = new MemBuf_2( vm, size * 2); break;
-      case 3: mb = new MemBuf_3( vm, size * 3); break;
-      case 4: mb = new MemBuf_4( vm, size * 4); break;
+      case 1: mb = new MemBuf_1( vm, (uint32) size ); break;
+      case 2: mb = new MemBuf_2( vm, (uint32) size * 2); break;
+      case 3: mb = new MemBuf_3( vm, (uint32) size * 3); break;
+      case 4: mb = new MemBuf_4( vm, (uint32) size * 4); break;
    }
    fassert( mb != 0 );
    vm->retval( mb );
